restructure app client
